fix: change int projectId to Long

+++ src/main/java/com/construtech/buildsphere/platform/documents/domain/model/aggregates/Document.java
@@ -24,13 +24,13 @@ public class Document extends AuditableAbstractAggregateRoot<Document> {
 
 
     public Document() {
-        this.project = new Project(0);
+        this.project = new Project(0L);
         this.name = "";
         this.description = "";
         this.fileType = "";
     }
 
-    public Document(int project, String name, String description, String fileType){
+    public Document(Long project, String name, String description, String fileType){
         this();
         this.project = new Project(project);
         this.name = name;
@@ -42,7 +42,7 @@ public class Document extends AuditableAbstractAggregateRoot<Document> {
         this.name = command.name();
         this.description = command.description();
         this.fileType = command.fileType();
-        this.project = new Project(command.project());
+        this.project = new Project((long) command.project());
     }
 
     public Document updateInformation(String name, String description, String fileType){
@@ -52,7 +52,7 @@ public class Document extends AuditableAbstractAggregateRoot<Document> {
         return this;
     }
 
-    public int getProjectId(){
+    public Long getProjectId(){
         return project.projectEnt();
     }
