+++ src/main/java/com/construtech/buildsphere/platform/documents/domain/model/commands/CreateDocumentCommand.java
@@ -1,4 +1,4 @@
 package com.construtech.buildsphere.platform.documents.domain.model.commands;
 
-public record CreateDocumentCommand(String name, String description, String fileType, int project) {
+public record CreateDocumentCommand(String name, String description, String fileType, Long project) {
 }
